[Store] Split sharded tenant quota headers
