+++ mooncake-store/include/master_service.h
@@ -29,7 +29,7 @@
 #include "master_metric_manager.h"
 #include "mutex.h"
 #include "segment.h"
-#include "tenant_quota.h"
+#include "tenant_quota_sharded.h"
 #include "tenant_quota_policy_store.h"
 #include "types.h"
 #include "master_config.h"
